break large page sections into components
