+++ components/FigImage.module.css
@@ -1,0 +1,24 @@
+.figure {
+  margin: 0 0 1rem;
+}
+
+.caption {
+  font-size: 1.25rem;
+  font-weight: 300;
+}
+
+.sm{
+  font-size: 75%;
+  line-height: 1;
+}
+
+.link {
+  color: inherit;
+  text-decoration: underline;
+}
+
+.figureImg {
+  margin: 0;
+  max-width: 100%;
+  height: auto;
+}
