+++ components/ImgCard.module.css
@@ -1,0 +1,28 @@
+.link {
+  color: inherit;
+  text-decoration: none;
+}
+
+.card {
+  height: 100%;
+  background-color: rgba(255,255,255,0.8);
+  border-radius: .375rem;
+  overflow: hidden;
+}
+
+.card:hover {
+  color: #ddd;
+  background-color: #222;
+}
+
+.img {
+  display: block;
+  width: 100%;
+}
+
+.p {
+  font-size: 1.25rem;
+  line-height: 1.2;
+  text-align: center;
+  font-weight: 500;
+}
