+++ css/styles.css
@@ -6,12 +6,12 @@ body {
 
   /* https://css-pattern.com/ */
   background:
-    conic-gradient(from 75deg,var(--c1)   15deg ,var(--c2) 0 30deg ,#0000 0 180deg,
-                              var(--c2) 0 195deg,var(--c1) 0 210deg,#0000 0) 
-       calc(0.5*var(--s)) calc(0.5*var(--s)/0.577),
-    conic-gradient(var(--c1)   30deg ,var(--c3) 0 75deg ,var(--c1) 0 90deg, var(--c2) 0 105deg,
-                   var(--c3) 0 150deg,var(--c2) 0 180deg,var(--c3) 0 210deg,var(--c1) 0 256deg,
-                   var(--c2) 0 270deg,var(--c1) 0 286deg,var(--c2) 0 331deg,var(--c3) 0);
+    conic-gradient( from 75deg,var(--c1)   15deg ,var(--c2) 0 30deg ,#0000 0 180deg,
+                    var(--c2) 0 195deg,var(--c1) 0 210deg,#0000 0) 
+    calc(0.5*var(--s)) calc(0.5*var(--s)/0.577),
+    conic-gradient( var(--c1)   30deg ,var(--c3) 0 75deg ,var(--c1) 0 90deg, var(--c2) 0 105deg,
+                    var(--c3) 0 150deg,var(--c2) 0 180deg,var(--c3) 0 210deg,var(--c1) 0 256deg,
+                    var(--c2) 0 270deg,var(--c1) 0 286deg,var(--c2) 0 331deg,var(--c3) 0);
   background-size: var(--s) calc(var(--s)/0.577);
   /* 0.577=tan(30deg)*/
