+++ pages/image/[id].js
@@ -1,43 +1,7 @@
 import Head from 'next/head'
-import {useRouter} from 'next/router'
+import FigImage from '../../components/FigImage'
 import {getImageById, getImages} from "../../util/database"
 import styles from "./image.module.css"
-
-
-const Image = ({image}) => {
-  const router = useRouter();
-  const getLink = (path) => `${router.basePath}${path}`;
-
-  return (
-    <main>
-      <Head>
-        <title>{image.title}</title>
-      </Head>
-
-      <section className={styles.container}>
-        <h1 className={styles.heading}>{image.title}</h1>
-        <figure className={styles.figure}>
-          <img src={getLink(`/${image.img}.jpg`)} className={styles.figureImg} alt={image.caption} />
-          <figcaption className={styles.caption}>
-            {!image.original ? " " :
-              <div className={styles.sm}>
-                <small>
-                  <a href={image.original}>Original image here</a>
-                </small>
-              </div>
-            }
-            <div className="text-center">{image.caption}</div>
-          </figcaption>
-        </figure>
-        {!image.story ? "" :
-          <div className={styles.story} dangerouslySetInnerHTML={{__html: image.story}}/>
-        }
-
-      </section>
-
-    </main>
-  );
-}
 
 export async function getStaticPaths() {
   const images = await getImages();
@@ -62,6 +26,28 @@ export async function getStaticProps({params}) {
   return {
     props: {image}
   };
+}
+
+const Image = ({image}) => {
+  return (
+    <main>
+      <Head>
+        <title>{image.title}</title>
+      </Head>
+
+      <section className={styles.container}>
+        <h1 className={styles.heading}>{image.title}</h1>
+        <FigImage 
+          img={image.img}
+          caption={image.caption}
+          original={image.original}
+        />
+        {!image.story ? "" :
+          <div className={styles.story} dangerouslySetInnerHTML={{__html: image.story}}/>
+        }
+      </section>
+    </main>
+  );
 }
 
 export default Image;
