+++ pages/image/image.module.css
@@ -14,31 +14,9 @@
   line-height: 1.2;
 }
 
-.figure {
-  margin: 0 0 1rem;
-}
-
-.caption {
-  font-size: 1.25rem;
-  font-weight: 300;
-}
-
-.sm{
-  font-size: 75%;
-  line-height: 1;
-}
-
-.link {
-  color: inherit;
-  text-decoration: underline;
-}
-
-.figureImg {
-  margin: 0;
-  max-width: 100%;
-  height: auto;
-}
-
 .story {
   padding: 2%
+}
+.story a {
+  color: inherit;
 }
